fix(#1211): keep app bridge within block boundary
